reception-breakthrough/requirements.txt
pytest>=8.0
pytest-cov>=5.0
sqlite-utils>=3.36
PyYAML>=6.0
